# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel OpenApi Demo Documentation",
        "description": "L5 Swagger OpenApi description",
        "contact": {
            "email": "admin@admin.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "Demo API Server"
        }
    ],
    "paths": {
        "/api/accounts": {
            "post": {
                "tags": [
                    "Account"
                ],
                "summary": "Create new account",
                "operationId": "create-account",
                "requestBody": {
                    "description": "Pass account details",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "type",
                                    "currency"
                                ],
                                "properties": {
                                    "type": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "cash"
                                    },
                                    "currency": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "usd"
                                    },
                                    "balance": {
                                        "type": "integer",
                                        "format": "integer",
                                        "example": "10000"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Parameter(s) missing",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}